backend/llm/requirements.txt
openai==1.51.2
python-dotenv==1.0.1